# tests/Swagger_updated_20-10-2017.yaml
swagger: '2.0'
info:
  version: '2.0'
  title: "API Project I'm Ready"
  description: "API Project I'm Ready"
# basePath: "/"
# host: "imready.ml"
...
produces:
  - "application/json"
schemes:
 - "https"
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "Authorization"
    in: "header"
...
definitions:
  PatchRequest:
    type: "array"
    items:
      $ref: "#/definitions/PatchDocument"
  PatchDocument:
    description: "A JSONPatch document as defined by RFC 6902"
    required:
    - "op"
    - "path"
    properties:
     op:
      type: "string"
      description: "The operation to be performed"
      enum:
      - "increment"
      - "replace"
     path:
      type: "string"
      description: "The key to be used within the operations"
     value:
      type: "object"
      description: "The value to be used within the operations"
  Clients:
    type: "object"
    properties:
      id:
        description: "ID of the user"
        type: "string"
      name:
        description: "Name of the user"
        type: "string"
      points:
        description: "Amount of points the user has"
        type: "integer"
        format: "int64"
      role:
        description: "Role of client"
        type: "string"
        enum:
        - "admin"
        - "client"
        - "caregiver"
        - "family"
      messages:
        description: "All messages this user has sent or received"
        type: "array"
        items: {
          $ref: "#/definitions/Messages"
        }
      family:
        description: "All family members invited by this user"
        type: "array"
        items: {
          $ref: "#/definitions/Family"
        }
      appointments:
        description: "All appointments of the user"
        type: "array"
        items: {
          $ref: "#/definitions/Appointments"
        }
      caregivers:
        description: "Caregivers of the client"
        type: "array"
        items: {
          $ref: "#/definitions/Caregivers"
        }
      blocks:
        description: "Personal building blocks"
        type: "array"
        items: {
          $ref: "#/definitions/Blocks"
        }
  Caregivers:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
        description: "Name of the user"
      role:
        description: "Role of client"
        type: "string"
        enum:
        - "admin"
        - "client"
        - "caregiver"
        - "family"
      messages:
        description: "All messages this user has sent or received"
        type: "array"
        items: {
          $ref: "#/definitions/Messages"
        }        
      appointments:
        description: "All appointments of the user"
        type: "array"
        items: {
          $ref: "#/definitions/Appointments"
        }
  EntityModel:
    type: "object"
    properties:
      id:
        type: "string"
  Family:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
        description: "Name of the family member"
      role:
        description: "Role of client"
        type: "string"
        enum:
        - "admin"
        - "client"
        - "caregiver"
        - "family"        
      messages:
        description: "All messages this user has sent or received"
        type: "array"
        items: {
          $ref: "#/definitions/Messages"
        }
...
  Blocks:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        description: "Name of the block"
        type: "string"
      description:
        description: "description of the block"
        type: "string"
      imageURL:
        description: "Path to the image"
        type: "string"
      activities:
        description: "Activities in the building block"
        type: "array"
        items: {
          $ref: "#/definitions/Activities"
        }
  Activities:
    type: "object"
    properties:
      id:
        description: "ID of the activity"
        type: "string"
      name:
        description: "Name of the activity"
        type: "string"
      description:
        description: "Description of the activity"
        type: "string"
...
  Appointments:
    type: "object"
    properties:
      id:
        description: "ID of the appointment"
        type: "string"
      datetime:
        description: "Deadline of the appointment"
        type: "string"
        format: "datetime"
      description:
        description: "Description of the appointment"
        type: "string"    
...
  Messages:
    type: "object"
    properties:
      id:
        description: "ID of the message"
        type: "string"
      senderId:
        description: "ID of the sender of the message"
        type: "string"
      receiverId:
        description: "ID of the receiver of the message"
        type: "string"
      message:
        description: "Content of the message"
        type: "string"
      datetime:
        description: "Time the message was sent"
        type: "string"
        format: "datetime"
      read:
        description: "The read status of the message"
        type: "boolean"
...
  Credentials:
    type: "object"
    properties:
      username: 
        description: "Username"
        type: "string"
      password:
        description: "Password"
        type: "string"
...
  ErrorResponses:
    type: "object"
    properties:
      code:
        description: "Error code"
        type: "integer"
        format: "int64"
      message:
        description: "Error message"
        type: "string"        
paths:
...
  /clients/{id}:
    get:
      summary: "Find client by ID"
      description: "Returns a single client"
      tags:
      - "general"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of client to return"
        required: true
        type: "string"
      responses:
        200:
          description: "succesfully found the client by ID"
          schema:
            $ref: "#/definitions/Clients"
        400:
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Client not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    put:
      summary: "Update client"
      description: "Update client with specified ID"
      tags:
      - "general"
      consumes:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of client to be updated"
        required: true
        type: "string"
      - name: "Credentials"
        in: "body"
        description: "Credentials of the user"
        required: true
        schema:
          $ref: "#/definitions/Credentials"        
      responses:
        200:
          description: "succesfully updated client"
        400:
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Client not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    patch:
      summary: "Update field within client"
      description: "Update a single field within a client, for example points" #{ "op": "replace", "path": "/points", "value": "2" }
      tags:
      - "general"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of client to be updated"
        required: true
        type: "string"
      - name: "body"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/PatchRequest"
      responses:
        200:
          description: "succesfully updated client"
        400:
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Client not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    delete:
      summary: "Delete client"
      description: "Delete client with specified ID"
      tags:
      - "general"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of client to be deleted"
        required: true
        type: "string"
      responses:
        204:
          description: "succesfully deleted client"
        400:
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Client not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []
...
  /caregivers/{id}:
    get:
      summary: "Get caregiver"
      description: "Returns a caregiver that fits the specified arguments."
      tags:
      - "general"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of caregiver to return"
        required: true
        type: "string"
      responses:
        200:
          description: "succesful operation"
          schema:
            $ref: "#/definitions/Caregivers"
        400:
          description: "Invalid parameters"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Caregiver not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    put:
      summary: "Update caregiver"
      description: "Update caregiver with specified ID"
      tags:
      - "general"
      consumes: 
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of caregiver to be updated"
        required: true
        type: "string"
      - name: "Credentials"
        in: "body"
        description: "Credentials of the user"
        required: true
        schema:
          $ref: "#/definitions/Credentials"
      responses:
        200:
          description: "succesfully updated caregiver"
        400:
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Caregiver not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    delete:
      summary: "Delete caregiver"
      description: "Delete caregiver with specified ID"
      tags:
      - "general"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the caregiver to deleted"
        required: true
        type: "string"
      responses:
        204:
          description: "succesfully deleted caregiver"
        400:
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Caregiver not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []
...
  /clients/{id}/caregivers:
    get:
      summary: "Get caregivers of a single client"
      description: "Returns a list of caregivers from the database that fits the specified client ID."
      tags:
      - "connections"
      parameters:
      - name: "id"
        in: "path"
        description: "The client's ID which caregivers to return"
        required: true
        type: "string"
      - name: "count"
        in: "query"
        description: "The maximum number of caregivers to return."
        required: false
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "succesful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Caregivers"
        400:
          description: "Invalid parameters"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    post:
      summary: "Bind caregiver to client"
      description: "Bind new caregiver to client"
      tags:
      - "connections"
      consumes:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Client's ID to modify"
        required: true
        type: "string"
      - name: "caregiver"
        in: "body"
        description: "ID of the caregiver"
        required: true
        schema:
          $ref: "#/definitions/EntityModel"
      responses:
        201:
          description: "Succesfully bound caregiver to client"
        400:
          description: "Invalid parameters"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []
  /clients/{clientId}/caregivers/{caregiverId}:
    get:
      summary: "Get a single client's caregiver with ID"
      description: "Returns a single caregiver with specified userID and caregiverID"
      tags:
        - "connections"
      parameters:
      - name: "clientId"
        in: "path"
        description: "ID of client to be returned"
        required: true
        type: "string"
      - name: "caregiverId"
        in: "path"
        description: "ID of the client's caregiver to be returned"
        required: true
        type: "string"
      responses:
        200:
          description: "succesfully found the caregigver by ID"
          schema:
            $ref: "#/definitions/Caregivers"
        400:
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Client or caregiver not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    put:
      summary: "Update client's caregiver"
      description: "Update caregiver with specified userID and caregiverID"
      tags:
      - "connections"
      consumes:
      - "application/json"
      parameters:
      - name: "clientId"
        in: "path"
        description: "ID of client to be updated"
        required: true
        type: "string"
      - name: "caregiverId"
        in: "path"
        description: "ID of the client's caregiver to be updated"
        required: true
        type: "string"
      - name: "caregiver"
        in: "body"
        description: "Updated caregiver object"
        required: true
        schema:
          $ref: "#/definitions/EntityModel"
      responses:
        200:
          description: "Succesfully updated client"
        400:
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Client or caregiver not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    delete:
      summary: "Delete client's caregiver"
      description: "Delete client's caregiver with specified ID"
      tags:
      - "connections"
      parameters:
      - name: "clientId"
        in: "path"
        description: "ID of client to be updated"
        required: true
        type: "string"
      - name: "caregiverId"
        in: "path"
        description: "ID of the client's caregiver to be deleted"
        required: true
        type: "string"
      responses:
        204:
          description: "Succesfully updated client"
        400:
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Client not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []
...
host: virtserver.swaggerhub.com
basePath: /InhollandCloudAPI/ProjectAPI/2.0